Tell Zeitwerk to ignore `app/overrides`
Rails 6 uses Zeitwerk by default, which expects each file under an
autoloaded path to define its corresponding constant. This causes
issues with `app/overrides`, so we need to tell Zeitwerk to ignore
`app/overrides` in the main app and all Railties.

Removing the `app/overrides` paths from `eager_load_paths` is the
right first step, but it's still not accounting for the fact that
Zeitwerk will still run its checks on any paths that are being
autoloaded via Zeitwerk (see [1]), so we need to also `ignore` any
`app/overrides` paths if we want Zeitwerk not to complain.

[1]: https://github.com/fxn/zeitwerk/blob/47fc4c82096dc4e55041cd1c6e61c3492876261b/lib/zeitwerk/kernel.rb#L13